+++ classes/Relations/Reference/Builder.php
@@ -149,15 +149,10 @@ trait Builder
      * @param string $localKeyProperty
      * @return $this
      * @note Passing a Property instance to setLocalKey is deprecated
-     * @noinspection PhpMissingParamTypeInspection
      */
-    public function setLocalKey($localKeyProperty): self
+    public function setLocalKey(string $localKeyProperty): self
     {
-        if (!$localKeyProperty instanceof Property) {
-            $localKeyProperty = $this->property($this->localClass, $localKeyProperty);
-        }
-
-        $this->localKeyProperty = $localKeyProperty;
+        $this->localKeyProperty = $this->property($this->localClass, $localKeyProperty);
 
         return $this;
     }
@@ -179,16 +174,6 @@ trait Builder
     /**
      * @param string $remoteKeyColumn
      * @return $this
-     * @deprecated Use setRemoteKeyColumn() instead
-     */
-    public function setRemoteKeyString(string $remoteKeyColumn): self
-    {
-        return $this->setRemoteKeyColumn($remoteKeyColumn);
-    }
-
-    /**
-     * @param string $remoteKeyColumn
-     * @return $this
      */
     public function setRemoteKeyColumn(string $remoteKeyColumn): self
     {
@@ -202,15 +187,10 @@ trait Builder
      * @param string $remoteKeyProperty
      * @return $this
      * @note Passing a Property instance to setRemoteKey is deprecated
-     * @noinspection PhpMissingParamTypeInspection
      */
-    public function setRemoteKey($remoteKeyProperty): self
+    public function setRemoteKey(string $remoteKeyProperty): self
     {
-        if (!$remoteKeyProperty instanceof Property) {
-            $remoteKeyProperty = $this->property($this->remoteClass, $remoteKeyProperty);
-        }
-
-        $this->remoteKeyProperty = $remoteKeyProperty;
+        $this->remoteKeyProperty = $this->property($this->remoteClass, $remoteKeyProperty);
         $this->remoteKeyColumn   = $this->policy->column($this->remoteKeyProperty->name());
 
         return $this;